Break up sensor_definition.py (#10181)
break up sensor_definition.py
