+++ python_modules/dagster/dagster/_core/definitions/__init__.py
@@ -85,7 +85,6 @@ from .schedule_definition import (
     ScheduleExecutionContext,
 )
 from .sensor_definition import (
-    AssetSensorDefinition,
     DefaultSensorStatus,
     SensorDefinition,
     SensorEvaluationContext,
